refactor: transfer SignUpRedirect and AccountMissing to components package
