+++ app/src/main/java/com/immortalidiot/rutlead/MainActivity.kt
@@ -11,7 +11,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             RUTLeadTheme {
-                // A surface container using the 'background' color from the theme
+                LoginDesign()
             }
         }
     }
