+++ app/src/main/java/com/immortalidiot/rutlead/components/SignUpRedirect.kt
@@ -1,4 +1,4 @@
-package com.immortalidiot.rutlead.fields
+package com.immortalidiot.rutlead.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.padding
